massive clean, new real test start

+++ real_tests/tests/ft_printf/08_ft_printf_print_f_0.c
@@ -6,7 +6,7 @@
 /*   By: ntoniolo <ntoniolo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/25 22:15:46 by ntoniolo          #+#    #+#             */
-/*   Updated: 2017/11/25 22:35:18 by ntoniolo         ###   ########.fr       */
+/*   Updated: 2017/11/26 14:52:53 by ntoniolo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,16 +15,15 @@
 int		ft_printf_print_f_0(void)
 {
 	char	buffer[UT_SIZE + 1];
-	char 	buffer2[UT_SIZE + 1];
+	char	buffer2[UT_SIZE + 1];
 	int		ret;
 	int		ret2;
 
 	ft_bzero(buffer, UT_SIZE);
 	ft_bzero(buffer2, UT_SIZE);
-
-	ret = ft_printf("[%f]", (float)0);
+	ret = ft_printf("[%f]", 0.0);
 	read(get_fd_out(), buffer, UT_SIZE);
-	ret2 = dprintf(1, "[%f]", (float)0);
+	ret2 = dprintf(1, "[%f]", 0.0);
 	read(get_fd_out(), buffer2, UT_SIZE);
 	if (!ft_strcmp(buffer2, buffer) || ret != ret2)
 		return (0);
